pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "freeread"
version = "0.1.0"
description = "Read paywalled articles in your terminal"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "readability-lxml>=0.8.1",
    "rich>=13.0.0",
    "html2text>=2024.2.26",
]

[project.scripts]
freeread = "freeread:main"

[tool.hatch.build.targets.wheel]
packages = ["freeread.py"]
